move classes to websocket folder
