+++ zeppelin-client/src/main/java/org/apache/zeppelin/client/websocket/MessageHandler.java
@@ -15,30 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.zeppelin.client;
+
+package org.apache.zeppelin.client.websocket;
 
 /**
- * MessageHandler for each statement.
+ * Interface of how to handle websocket message sent from ZeppelinServer.
  */
-public interface StatementMessageHandler {
+public interface MessageHandler {
 
-  /**
-   * Invoked when there's new statement output appended.
-   *
-   * @param statementId
-   * @param index
-   * @param output
-   */
-  void onStatementAppendOutput(String statementId, int index, String output);
-
-  /**
-   * Invoked when statement's output is updated.
-   *
-   * @param statementId
-   * @param index
-   * @param type
-   * @param output
-   */
-  void onStatementUpdateOutput(String statementId, int index, String type, String output);
+  void onMessage(String msg);
 
 }
