+++ zeppelin-client/src/main/java/org/apache/zeppelin/client/websocket/SimpleMessageHandler.java
@@ -15,14 +15,25 @@
  * limitations under the License.
  */
 
+package org.apache.zeppelin.client.websocket;
 
-package org.apache.zeppelin.client;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
- * Interface of how to handle websocket message sent from ZeppelinServer.
+ * Simple implementation of AbstractMessageHandler which only print output.
  */
-public interface MessageHandler {
+public class SimpleMessageHandler extends AbstractMessageHandler {
 
-  void onMessage(String msg);
+  private static final Logger LOGGER = LoggerFactory.getLogger(SimpleMessageHandler.class);
 
+  @Override
+  public void onStatementAppendOutput(String statementId, int index, String output) {
+    LOGGER.info("Append output, data: {}", output);
+  }
+
+  @Override
+  public void onStatementUpdateOutput(String statementId, int index, String type, String output) {
+    LOGGER.info("Update output, type: {}, data: {}", type, output);
+  }
 }
